Disposition: Factor out transports into their own module.
