+++ opengever/disposition/testing.py
@@ -1,8 +1,8 @@
-from opengever.disposition.delivery import BaseTransport
 from opengever.disposition.delivery import DeliveryScheduler
 from opengever.disposition.interfaces import IDisposition
 from opengever.disposition.interfaces import IFilesystemTransportSettings
 from opengever.disposition.interfaces import ISIPTransport
+from opengever.disposition.transports import BaseTransport
 from opengever.testing import IntegrationTestCase
 from opengever.testing.helpers import CapturingLogHandler
 from plone.registry.interfaces import IRegistry
