+++ opengever/disposition/transports/__init__.py
@@ -1,0 +1,2 @@
+from .base import BaseTransport  # noqa
+from .filesystem import FilesystemTransport  # noqa
